fixed double number sqlite bug

+++ app/src/main/java/augarte/sendo/dataModel/Measurement.kt
@@ -5,7 +5,7 @@ import java.util.*
 class Measurement {
     var id: Int? = null
     var type: MeasureType? = null
-    var value: Int? = null
+    var value: Double? = null
     var date: Date? = null
     var createdBy: User? = null
     var createDate: Date? = null
